data_pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.5
hdbscan>=0.8.33
chromadb>=0.4.0
sentence-transformers>=2.2.2
tqdm>=4.66.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
plotly>=5.18.0
